app/stockAI-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
boto3>=1.26.0
requests>=2.28.0
python-dotenv>=0.19.0
requests>=2.28.0
openai==1.108.1